fixing tutorial touch view

+++ Utopia/Classes/Battle/Skills/SkillEnergize.m
@@ -277,7 +277,7 @@ static const NSInteger kBatteryOrbsMaxSearchIterations = 256;
   OrbSwipeLayer* layer = self.battleLayer.orbLayer.swipeLayer;
   OrbBgdLayer* bgdLayer = self.battleLayer.orbLayer.bgdLayer;
   
-  NSInteger usedUpOrbCount = 0;
+  int usedUpOrbCount = 0;
   NSMutableSet* clonedOrbs = [NSMutableSet set];
   
   for (NSInteger column = 0; column < layout.numColumns; ++column)
